+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Added limited support for tables. 
 - Added support for custom labels and start indices for enumerate environments. 
+- Added support for arbitrary macro usage in labels. 
 
 
 # v0.2.2, 2026-04-12